% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{ls.nodes}
\alias{ls.nodes}
\title{Gets a data frame of `node` metadata}
\usage{
ls.nodes(filters = NULL)
}
\arguments{
\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A data frame of node metadata
}
\description{
Nodes are the physical devices deployed to collect observations.
The are comprised of multiple sensors and are grouped by
projects.
}
