% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{ls.observations}
\alias{ls.observations}
\title{Gets a data frame of `obserations` data.}
\usage{
ls.observations(filters = NULL)
}
\arguments{
\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A data frame of observation data
}
\description{
Observation data are the environmental measurements made
by the sensors. Data listed here is more or less tuned
and trustworthy.
}
