% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{ls.projects}
\alias{ls.projects}
\title{Gets a data frame of `project` metadata}
\usage{
ls.projects(filters = NULL)
}
\arguments{
\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A data frame of project metadata
}
\description{
Projects are the highest entity in the hierarchy of the Array of
Things system. They are generally ambiguous geographic regions used
to roughly group nodes.
}
