% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{ls.raw_observations}
\alias{ls.raw_observations}
\title{Gets a data frame of `raw observations` data.}
\usage{
ls.raw_observations(filters = NULL)
}
\arguments{
\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A data frame of project metadata
}
\description{
Raw observation data are the environmental measurements made
by the sensors. Data listed here can be both tuned and not-yet
tuned data. The values of `raw` are the analog readings made by
the sensor; `hrf` (human readable format) are the clean/trusted
values -- these are typically null.
}
