% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{ls.sensors}
\alias{ls.sensors}
\title{Gets a data frame of `sensor` metadata}
\usage{
ls.sensors(filters = NULL)
}
\arguments{
\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A data frame of sensor metadata
}
\description{
Sensors are the physical boards inside the nodes that record
the observations. Sensors are in various states of being tuned
and therefor some of their observations are considered to be
experimental. Trustworthy data is listed under the _observations_
endpoint and experimental data is under _raw-observations_.
}
