% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{stat.node}
\alias{stat.node}
\title{Gets the details for a single `node` metadata record}
\usage{
stat.node(vsn, filters = NULL)
}
\arguments{
\item{vsn}{- The node"s unique identifier}

\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A list representing the node
}
\description{
Nodes are the physical devices deployed to collect observations.
The are comprised of multiple sensors and are grouped by
projects.
}
