% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{stat.project}
\alias{stat.project}
\title{Gets the details for a single `project` metadata record}
\usage{
stat.project(slug, filters = NULL)
}
\arguments{
\item{slug}{- The project"s unique identifier}

\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A list representing the project
}
\description{
Projects are the highest entity in the hierarchy of the Array of
Things system. They are generally ambiguous geographic regions used
to roughly group nodes.
}
