% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AotClient.R
\name{stat.sensor}
\alias{stat.sensor}
\title{Gets the details for a single `sensor` metadata record}
\usage{
stat.sensor(path, filters = NULL)
}
\arguments{
\item{path}{- The sensor"s unique identifier}

\item{filters}{- A list of tuples to create filters/query params}
}
\value{
A list representing the sensor
}
\description{
Sensors are the physical boards inside the nodes that record
the observations. Sensors are in various states of being tuned
and therefor some of their observations are considered to be
experimental. Trustworthy data is listed under the _observations_
endpoint and experimental data is under _raw-observations_.
}
